requirements.txt
python-telegram-bot==20.8
anthropic==0.18.1
openai==1.68.0
python-dotenv==1.0.1
moviepy==1.0.3
ffmpeg-python==0.2.0
numpy>=2.0.2
tqdm==4.67.1
requests==2.31.0
aiohttp==3.9.3
Pillow==10.2.0
httpx>=0.24.1
httpcore>=0.17.0
yt-dlp==2025.1.26
youtube-transcript-api==0.6.2
google-api-python-client==2.118.0
